Rename printoptions package to print
