+++ java/client/src/org/openqa/selenium/PrintsPage.java
@@ -16,7 +16,7 @@
 // under the License.
 package org.openqa.selenium;
 
-import org.openqa.selenium.printoptions.PrintOptions;
+import org.openqa.selenium.print.PrintOptions;
 
 public interface PrintsPage {
   Pdf print(PrintOptions printOptions) throws WebDriverException;
