+++ java/client/src/org/openqa/selenium/print/BUILD.bazel
@@ -10,7 +10,6 @@ java_library(
         "//java/client/test/org/openqa/selenium:__subpackages__",
         "//java/client/test/org/openqa/selenium/build:__pkg__",
         "//java/client/test/org/openqa/selenium/testing/drivers:__pkg__",
-        "//java/client/src/org/openqa/selenium:__subpackages__",
     ],
     deps = [
         "//java/client/src/org/openqa/selenium:core",
