+++ java/client/src/org/openqa/selenium/print/PageMargin.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.printoptions.print;
+package org.openqa.selenium.print;
 
 import org.openqa.selenium.internal.Require;
