+++ java/client/src/org/openqa/selenium/print/PageSize.java
@@ -15,10 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.printoptions.print;
+package org.openqa.selenium.print;
 
-import java.util.HashMap;
-import java.util.Map;
 
 public class PageSize {
