--- java/client/src/org/openqa/selenium/printoptions/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@rules_jvm_external//:defs.bzl", "artifact")
-load("//java:defs.bzl", "java_library")
-
-java_library(
-    name = "printoptions",
-    srcs = glob(["*.java"]),
-    visibility = [
-        "//java/client/src/org/openqa/selenium:__pkg__",
-        "//java/client/src/org/openqa/selenium/remote:__pkg__",
-        "//java/client/test/org/openqa/selenium:__subpackages__",
-        "//java/client/test/org/openqa/selenium/build:__pkg__",
-        "//java/client/test/org/openqa/selenium/testing/drivers:__pkg__",
-    ],
-    deps = [
-        "//java/client/src/org/openqa/selenium/printoptions/print",
-    ]
-)
