+++ java/client/src/org/openqa/selenium/remote/DriverCommand.java
@@ -25,7 +25,7 @@ import org.openqa.selenium.Dimension;
 import org.openqa.selenium.Point;
 import org.openqa.selenium.WindowType;
 import org.openqa.selenium.interactions.Sequence;
-import org.openqa.selenium.printoptions.PrintOptions;
+import org.openqa.selenium.print.PrintOptions;
 
 import java.time.Duration;
 import java.util.Collection;
