+++ java/client/src/org/openqa/selenium/remote/RemoteWebDriver.java
@@ -55,7 +55,8 @@ import org.openqa.selenium.logging.LoggingHandler;
 import org.openqa.selenium.logging.LoggingPreferences;
 import org.openqa.selenium.logging.Logs;
 import org.openqa.selenium.logging.NeedsLocalLogs;
-import org.openqa.selenium.printoptions.PrintOptions;
+import org.openqa.selenium.print.PageMargin;
+import org.openqa.selenium.print.PrintOptions;
 import org.openqa.selenium.remote.internal.WebElementToJsonConverter;
 import org.openqa.selenium.virtualauthenticator.Credential;
 import org.openqa.selenium.virtualauthenticator.HasVirtualAuthenticator;
