+++ java/client/test/org/openqa/selenium/chrome/BUILD.bazel
@@ -13,7 +13,6 @@ java_selenium_test_suite(
     ],
     deps = [
         "//java/client/src/org/openqa/selenium/chrome",
-        "//java/client/src/org/openqa/selenium/printoptions",
         "//java/client/src/org/openqa/selenium/remote",
         "//java/client/src/org/openqa/selenium/support",
         "//java/client/test/org/openqa/selenium/build",
