+++ java/client/test/org/openqa/selenium/print/BUILD.bazel
@@ -6,10 +6,9 @@ java_test_suite(
     size = "small",
     srcs = glob(["*Test.java", "print/*Test.java"]),
     deps = [
-        "//java/client/src/org/openqa/selenium/printoptions",
         "//java/client/test/org/openqa/selenium/testing:test-base",
         "//java/client/test/org/openqa/selenium/testing:annotations",
-        "//java/client/src/org/openqa/selenium/printoptions/print",
+        "//java/client/src/org/openqa/selenium/print:print",
         artifact("org.assertj:assertj-core"),
         artifact("junit:junit"),
     ]
